Refactor journal mode out of protocol and into implementation
