+++ apollo-ios/Sources/ApolloSQLite/SQLiteDatabase.swift
@@ -24,12 +24,6 @@ public protocol SQLiteDatabase {
   
   func clearDatabase(shouldVacuumOnClear: Bool) throws
 
-  func setJournalMode<T: RawRepresentable>(mode: T) throws where T.RawValue == String
-
-}
-
-public extension SQLiteDatabase {
-  func setJournalMode<T>(mode: T) throws where T : RawRepresentable, T.RawValue == String { }
 }
 
 public extension SQLiteDatabase {
