+++ apollo-ios/Sources/ApolloSQLite/SQLiteDotSwiftDatabase.swift
@@ -11,25 +11,6 @@ public final class SQLiteDotSwiftDatabase: SQLiteDatabase {
   private let keyColumn: SQLite.Expression<CacheKey>
   private let recordColumn: SQLite.Expression<String>
 
-  public enum JournalMode: String {
-    /// The rollback journal is deleted at the conclusion of each transaction. This is the default behaviour.
-    case delete = "DELETE"
-    /// Commits transactions by truncating the rollback journal to zero-length instead of deleting it.
-    case truncate = "TRUNCATE"
-    /// Prevents the rollback journal from being deleted at the end of each transaction. Instead, the header
-    /// of the journal is overwritten with zeros.
-    case persist = "PERSIST"
-    /// Stores the rollback journal in volatile RAM. This saves disk I/O but at the expense of database
-    /// safety and integrity.
-    case memory = "MEMORY"
-    /// Uses a write-ahead log instead of a rollback journal to implement transactions. The WAL journaling
-    /// mode is persistent; after being set it stays in effect across multiple database connections and after 
-    /// closing and reopening the database.
-    case wal = "WAL"
-    /// Disables the rollback journal completely
-    case off = "OFF"
-  }
-
   public init(fileURL: URL) throws {
     self.records = Table(Self.tableName)
     self.keyColumn = Expression<CacheKey>(Self.keyColumnName)
@@ -88,8 +69,8 @@ public final class SQLiteDotSwiftDatabase: SQLiteDatabase {
 
   /// Sets the journal mode for the current database.
   ///
-  /// - Parameter mode: Use ``JournalMode``.
-  public func setJournalMode<T>(mode: T) throws where T : RawRepresentable, T.RawValue == String {
+  /// - Parameter mode: The journal mode controls how the journal file is stored and processed.
+  public func setJournalMode(mode: JournalMode) throws {
     try self.db.run("PRAGMA journal_mode = \(mode.rawValue)")
   }
 }
